EDID with interlaced TV CEA PAL625 NTSC
And the most compatible 16 bit stereo AUDIO PCM formats to get Kodi running on a NVIDIA card with a new proprietary driver on Ubuntu on an old TV-set via a HDMI to VGA-adaptor that converts the picture further to RGB/composite.
I couldn't make this file with the edid-generator of this project as its current version ignores the interlace-tag. I however was able to create it with the Advantiv EDID-editor 0.97 (at the time of writing) at https://ez.analog.com/docs/DOC-2143.

I wonder whether this edid-generator also creates things on the CEA-tab. Will modelines on the VESA-tab support HDMI-audio?